ai-backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.0
asn1crypto==1.5.1
astroid==3.3.5
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.10
cryptography==43.0.1
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
filelock==3.16.1
frozenlist==1.4.1
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jiter==0.6.1
Mako==1.3.5
MarkupSafe==3.0.1
mccabe==0.7.0
multidict==6.1.0
nanoid==2.0.0
numpy==2.1.2
openai==1.51.2
packaging==24.1
passlib==1.7.4
pgvector==0.3.5
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
setuptools==75.1.0
six==1.16.0
slack_bolt==1.21.0
slack_sdk==3.33.1
sniffio==1.3.1
snowflake-connector-python==3.12.2
snowflake-sqlalchemy==1.6.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
sqlglot==25.28.0
starkbank-ecdsa==2.2.0
starlette==0.38.6
tabulate==0.9.0
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
wheel==0.44.0
yarl==1.15.5
